+++ domain/event-models/src/types/command.rs
@@ -62,7 +62,7 @@ impl ModifiablyDescribed for Command {
         self.description = description.to_string();
     }
 
-    fn add_to_description(&mut self, index: u32, addition: &str) {
+    fn add_to_description(&mut self, index: usize, addition: &str) {
         if self.description.is_empty() {
             self.set_description(addition);
         } else {
@@ -70,7 +70,7 @@ impl ModifiablyDescribed for Command {
         }
     }
 
-    fn delete_from_description(&mut self, index: u32) {
+    fn delete_from_description(&mut self, index: usize) {
         if !self.description.is_empty() {
             self.description.remove(index as usize);
         }
