+++ domain/event-models/src/types/mod.rs
@@ -51,11 +51,11 @@ pub trait Described: Named {
 
 pub trait ModifiablyDescribed: Described {
     fn set_description(&mut self, description: &str);
-    fn add_to_description(&mut self, index: u32, addition: &str);
-    fn delete_from_description(&mut self, index: u32);
+    fn add_to_description(&mut self, index: usize, addition: &str);
+    fn delete_from_description(&mut self, index: usize);
 }
 
-pub type LaneIndex = u32;
+pub type LaneIndex = usize;
 
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
 pub enum LaneId {
