+++ domain/event-models/src/types/read_model.rs
@@ -61,7 +61,7 @@ impl ModifiablyDescribed for ReadModel {
         self.description = description.to_string();
     }
 
-    fn add_to_description(&mut self, index: u32, addition: &str) {
+    fn add_to_description(&mut self, index: usize, addition: &str) {
         if self.description.is_empty() {
             self.set_description(addition);
         } else {
@@ -69,7 +69,7 @@ impl ModifiablyDescribed for ReadModel {
         }
     }
 
-    fn delete_from_description(&mut self, index: u32) {
+    fn delete_from_description(&mut self, index: usize) {
         if !self.description.is_empty() {
             self.description.remove(index as usize);
         }
